salesdb/requirements.txt
psycopg2-binary      
Faker                
pytest               
python-dotenv        
